move autodraws
